Add compression algorithm
This example reads a grayscale input image and generates a time-varying fractal heightmap as before. The compressImage function encodes the input image by adding the heightmap values to the pixel values of the input image. The decompressImage function reverses this process by subtracting the heightmap values from the pixel values of the compressed image.

Please note that this script assumes that originalImage and heightmap are BufferedImage objects. You will need to provide the appropriate image and heightmap data in your Main.java file to use the compression and decompression methods.